src/main/java/com/ukrposhta/lib/validator/EmailValidator.java
package com.ukrposhta.lib.validator;

import com.ukrposhta.lib.Email;
import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;

public class EmailValidator implements ConstraintValidator<Email, String> {
    private static final String EMAIL_VALIDATOR_PATTERN =
            "^\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$";

    @Override
    public boolean isValid(String email, ConstraintValidatorContext context) {
        return email != null && email.matches(EMAIL_VALIDATOR_PATTERN);
    }
}
